TableComparator
Table Comparator translated to Typescript. A new auxiliar type has been created  to ease the TableComparator code. The signature of one of the methods of Tablecomparator has changed so the controller has been updated properly.
TableComparator is partially documented.
